+++ src/layouts/DashboardLayout/index.jsx
@@ -1,33 +1,21 @@
 import { useEffect, useState } from "react";
-import { useDispatch, useSelector } from "react-redux";
+import { useSelector } from "react-redux";
 import { useNavigate } from "react-router-dom";
-
-import { fetchNotifications } from "~/store/actions/notificationActions";
 
 import NavigationBar from "./components/NavigationBar";
 
 function DashboardLayout({ children }) {
   const [sidebarOpen, setSidebarOpen] = useState(false);
-  const dispatch = useDispatch();
+
   const navigate = useNavigate();
 
   const {
-    data: { _id: memberId, token },
+    data: { token },
   } = useSelector((state) => state.user);
-
-  const {
-    data: { roomId },
-  } = useSelector((state) => state.chat);
 
   useEffect(() => {
     if (!token) navigate("/login");
   }, [token, navigate]);
-
-  useEffect(() => {
-    if (token) {
-      dispatch(fetchNotifications(memberId, roomId));
-    }
-  }, [dispatch, token, memberId, roomId]);
 
   return (
     <div className="min-h-screen">
